Crunch Day 4
Played with textures a lot. Built a very co-dependent system to track the player and trace back to which void they're interacting with. Would be happier with a delegate-inspired system but I wanted this working. Consider some thought on that if the current system becomes untnenable. Time is a precious resource at this point though.